% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Distance Matrix}
\usage{
distance_matrix(x_coords, y_coords)
}
\arguments{
\item{x_coords}{the x-coordinates of the graph nodes, as a vector.}

\item{y_coords}{the y-coordinates of the graph nodes, as a vector.}
}
\value{
a matrix D, where D[i,j] is the euclidean distance between node i and node j.
}
\description{
distance_matrix creates a distance matrix for the nodes of a graph, based on their euclidean distance.
}
\examples{
xs <- c(1,0,3)
ys <- c(1,2,4)
distance_matrix(xs,ys)

}
