% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{edge_set}
\alias{edge_set}
\title{Edges of a Graph from Adjacency Matrix}
\usage{
edge_set(adj_mat)
}
\arguments{
\item{adj_mat}{the adjacency matrix, A, for a graph.
If A[i,j] == 1, then nodes i and j are connected. If A[i,j] == 0, then they are not.}
}
\value{
an n x 2 matrix E, where n is the number of edges of the graph.
If there is a row with one element equal to i and one element equal to j, then there is an edge between nodes i and j.
}
\description{
edge_set converts an adjacency matrix for a graph into a matrix representing the edges of the graph.
Note that if a node is connected to itself, this edge will be ignored.
}
\examples{
test <- matrix(rep(0),nrow=4,ncol=4)
test[1,2] <- 1
test[2,1] <- 1
test[2,3] <- 1
test[3,2] <- 1
test[3,4] <- 1
test[4,3] <- 1
test[4,1] <- 1
test[1,4] <- 1
edge_set(test)

}
