% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{graph_plot}
\alias{graph_plot}
\title{Plot Graph of Function}
\usage{
graph_plot(x_coords, y_coords, adj_mat)
}
\arguments{
\item{x_coords}{the x-coordinates of the graph nodes, as a vector.}

\item{y_coords}{the y-coordinates of the graph nodes, as a vector.}

\item{adj_mat}{the adjacency matrix, A, for a graph.
If A[i,j] == 1, then nodes i and j are connected. If A[i,j] == 0, then they are not.}
}
\description{
graph_plot plots the nodes of a graph, and the links between them based on their adjacency matrix.
}
\examples{
xs <- c(1,1,2,2)
ys <- c(1,2,2,1)
test <- rbind(c(0,1,0,1),c(1,0,1,0),c(0,1,0,1),c(1,0,1,0))
plot_graph(xs,ys,test)

}
