% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prim}
\alias{prim}
\title{Minimum Spanning Tree using Prim's Algorithm}
\usage{
prim(x_coords, y_coords)
}
\arguments{
\item{x_coords}{the x-coordinates of the graph nodes, as a vector.}

\item{y_coords}{the y-coordinates of the graph nodes, as a vector.}
}
\value{
the adjacency matrix, A, for a minimimum spanning tree of the graph.
If A[i,j] == 1, then nodes i and j are connected. If A[i,j] == 0, then they are not.
}
\description{
prim finds a minimum spanning tree for the nodes of a graph, based on their euclidean distance.
Note that if there is more than one minimum spanning tree, only one will be returned.
}
\examples{
xs <- c(1,0,3)
ys <- c(1,2,4)
prim(xs,ys)

}
